# [Perf] Generate performance comparison report from one-line metrics summary.