This commit adds `scripts/perf_report.py`, which will generate a performance comparison summary report each time we run `./test.sh` to exercise the `run-all-client-server-perf-tests()` test-method.

- Change server-code slightly to generate a one-line summary in a canonical format. This is done under new `--perf-outfile` cmdline flag, and invoked by `./test.sh for run-all-client-server-perf-tests` method. (Manual execution of server program or executing other test-methods will -not- trigger this one-line summary output.)

- Add `scripts/perf_report.py` - parses the one-line summary generated in the file output under --perf-outfile flag, to generate the performance comparison report.

- Add `tests/pytests/perf_report_test.py`, to verify parsing logic.
- `test.sh`: Cleanup messages. Add execn-timing metrics.

There is a bit of 'handshake' between the one-line summary line generated by the server program, which is parsed by the perf_report.py script. This summary-line format handshake is verified by test-methods in the newly added perf_report_test.py pytest.

### Sample output:

As generated by the execution of `./test.sh run-all-client-server-perf-tests` in CI Linux jobs:

```
   **** Performance comparison for NumClients=5, NumOps=5000 (5 K) ****
+--------------------------+-------------------+----------+-------------------+----------+
|         Run-Type         | Server throughput | S:%-Drop | Client throughput | C:%-Drop |
+--------------------------+-------------------+----------+-------------------+----------+
|  Baseline - No logging   |    993 ops/sec    |   0.0    |  ~39.16 K ops/sec |   0.0    |
|   L3-logging (no LOC)    |    994 ops/sec    |   0.1    |  ~40.67 K ops/sec |   3.84   |
| L3-fast logging (no LOC) |    993 ops/sec    |   0.0    |  ~38.18 K ops/sec |  -2.51   |
|  L3-logging default LOC  |    993 ops/sec    |   0.0    |  ~38.94 K ops/sec |  -0.57   |
|    L3-logging LOC-ELF    |    994 ops/sec    |   0.1    |  ~41.19 K ops/sec |   5.18   |
+--------------------------+-------------------+----------+-------------------+----------+

```

### Timing messages added to test.sh

Show these output lines when running `run-all-client-server-perf-tests` in CI:

```
test.sh: Mon May 27 11:40:18 PDT 2024 Executing run-all-client-server-perf-tests ...

test.sh: Mon May 27 11:40:48 PDT 2024 Completed all client(s)-server communication test [ 0h 0m 30s ].
```